Cambiamenti TC per test Baita

+++ dual_core_completo/dual_core/cpu2_control_unit_test/timer_management.h
@@ -11,7 +11,7 @@
 extern Uint16 increment;
 
 extern Uint32 time_elapsed;
-extern Uint32 T_s;
+extern double T_s;
 
 extern bool brakeWhenSlow;
 extern bool brakeReg;
